+++ apx_sdk/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Added `VerificationMethod` to `CoreType` enum.
+
 ## [0.12.0] - 2025-04-24
 
 ### Added